+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+Feature:
+- Update `token.js` tests.
 
 ## [2.0.0] - 2019-05-20$